11/01/2024 review changes done

+++ src/utility/AsyncStorage.js
@@ -3,9 +3,15 @@ import EncryptedStorage from 'react-native-encrypted-storage';
 
 async function StoreAsyncData(key, body) {
   try {
+    const existingValue = await EncryptedStorage.getItem(key);
+
+    if (existingValue) {
+      return;
+    }
     await EncryptedStorage.setItem(key, JSON.stringify(body));
+    console.log('Value saved successfully');
   } catch (error) {
-    console.error('error occured', error);
+    console.error('Error occurred', error);
     Alert.alert('', 'Something Went Wrong');
   }
 }
